moved display_renderer function to streamlit
